# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Igloo API",
    "version": "0.1.0",
    "description": "HTTP API for the Igloo self-hosted media center. Most JSON endpoints return a shared envelope with error, optional message, and optional data fields."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin Igloo server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/tmdb/movies/in-theaters": {
      "get": {
        "tags": [
          "TMDB"
        ],
        "summary": "List movies currently in theaters",
        "operationId": "getMoviesInTheaters",
        "responses": {
          "200": {
            "$ref": "#/components/responses/TmdbTheaterMoviesResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/tmdb/movies/{id}": {
      "get": {
        "tags": [
          "TMDB"
        ],
        "summary": "Get TMDB movie details by TMDB ID",
        "operationId": "getMovieByTmdbID",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TmdbMovieResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/movies/{id}/tmdb-search": {
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "Search TMDB candidates for a local movie",
        "description": "Admin-only endpoint.",
        "operationId": "tmdbSearchMovies",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/TmdbSearchMoviesRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TmdbSearchResultsResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/music/stats": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "Get music library statistics",
        "operationId": "getMusicStats",
        "responses": {
          "200": {
            "$ref": "#/components/responses/MusicStatsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/music/tracks/": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "List tracks alphabetically",
        "operationId": "getTracksAlphabetical",
        "parameters": [
          {
            "$ref": "#/components/parameters/LimitQuery"
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TracksResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/music/tracks/shuffle": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "Get random tracks for shuffle playback",
        "operationId": "getShuffleTracks",
        "parameters": [
          {
            "$ref": "#/components/parameters/ShuffleLimitQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ShuffleTracksResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/music/tracks/details/{id}": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "Get track details",
        "operationId": "getTrackByID",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TrackDetailsResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/music/tracks/{id}/stream": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "Direct-stream an audio track",
        "operationId": "streamTrack",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          },
          {
            "$ref": "#/components/parameters/RangeHeader"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BinaryMediaResponse"
          },
          "206": {
            "$ref": "#/components/responses/PartialBinaryMediaResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/music/tracks/{id}/like": {
      "post": {
        "tags": [
          "Music"
        ],
        "summary": "Toggle current user's like for a track",
        "operationId": "toggleLikeTrack",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TrackLikeToggleResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/music/tracks/liked": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "List tracks liked by the current user",
        "operationId": "getLikedTracks",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/MusicPerPageQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/LikedTracksResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/music/tracks/liked-ids": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "List IDs of tracks liked by the current user",
        "operationId": "getLikedTrackIDsForUser",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LikedTrackIDsResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Server-side session cookie set by POST /api/auth/login."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
...
      "PageQuery": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
...
      "MusicPerPageQuery": {
        "name": "per_page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
...
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "ShuffleLimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      },
...
      "OffsetQuery": {
        "name": "offset",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
...
      "RangeHeader": {
        "name": "Range",
        "in": "header",
        "required": false,
        "description": "HTTP byte range for seeking.",
        "schema": {
          "type": "string",
          "example": "bytes=0-1048575"
        }
      }
    },
    "requestBodies": {
...
      "TmdbSearchMoviesRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TmdbSearchMoviesRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "TmdbTheaterMoviesResponse": {
        "description": "TMDB in-theaters response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TmdbTheaterMoviesEnvelope"
            }
          }
        }
      },
      "TmdbMovieResponse": {
        "description": "TMDB movie details response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TmdbMovieEnvelope"
            }
          }
        }
      },
      "TmdbSearchResultsResponse": {
        "description": "TMDB search results response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TmdbSearchResultsEnvelope"
            }
          }
        }
      },
...
      "BinaryMediaResponse": {
        "description": "Complete media file response.",
        "content": {
          "application/octet-stream": {
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
      "PartialBinaryMediaResponse": {
        "description": "Partial media file response for a byte range.",
        "headers": {
          "Content-Range": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/octet-stream": {
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
...
      "TracksResponse": {
        "description": "Paginated tracks response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TracksEnvelope"
            }
          }
        }
      },
      "ShuffleTracksResponse": {
        "description": "Shuffle tracks response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ShuffleTracksEnvelope"
            }
          }
        }
      },
      "TrackDetailsResponse": {
        "description": "Track detail response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TrackDetailsEnvelope"
            }
          }
        }
      },
      "TrackLikeToggleResponse": {
        "description": "Track like toggle response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TrackLikeToggleEnvelope"
            }
          }
        }
      },
      "LikedTracksResponse": {
        "description": "Liked tracks response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LikedTracksEnvelope"
            }
          }
        }
      },
      "LikedTrackIDsResponse": {
        "description": "Liked track IDs response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LikedTrackIDsEnvelope"
            }
          }
        }
      },
      "MusicStatsResponse": {
        "description": "Music statistics response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MusicStatsEnvelope"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Bad request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid session.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated user is not allowed to perform this action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": true
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "TmdbMovie": {
        "type": "object",
        "additionalProperties": true
      },
...
      "TmdbSearchMoviesRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "tmdb_id": {
            "type": "integer"
          }
        },
        "additionalProperties": false
      },
...
      "TmdbTheaterMoviesEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "movies"
                ],
                "properties": {
                  "movies": {
                    "type": "array",
                    "maxItems": 12,
                    "items": {
                      "$ref": "#/components/schemas/TheaterMovie"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "TmdbMovieEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "movie"
                ],
                "properties": {
                  "movie": {
                    "$ref": "#/components/schemas/TmdbMovie"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "TmdbSearchResultsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "results"
                ],
                "properties": {
                  "results": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TmdbSearchResult"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
...
      "Track": {
        "type": "object",
        "additionalProperties": true
      },
...
      "TracksEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "tracks",
                  "total",
                  "offset",
                  "limit",
                  "has_more"
                ],
                "properties": {
                  "tracks": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TrackListItem"
                    }
                  },
                  "total": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "offset": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "limit": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "has_more": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "ShuffleTracksEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "tracks"
                ],
                "properties": {
                  "tracks": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TrackListItem"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "TrackDetailsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "track"
                ],
                "properties": {
                  "track": {
                    "$ref": "#/components/schemas/Track"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "TrackLikeToggleEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "track_id",
                  "is_liked"
                ],
                "properties": {
                  "track_id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "is_liked": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
...
      "LikedTrackIDsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "liked_track_ids"
                ],
                "properties": {
                  "liked_track_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "MusicStatsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/MusicStats"
              }
            }
          }
        ]
      },
...
    }
  }
}